+++ src/main/java/io/github/devatherock/scheduler/KafkaLagCollector.java
@@ -1,10 +1,12 @@
 package io.github.devatherock.scheduler;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.Future;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
@@ -44,14 +46,14 @@ public class KafkaLagCollector {
     private static final String TAG_GROUP = "group";
     private static final String TAG_PARTITION = "partition";
     private static final String TAG_CLUSTER_NAME = "cluster_name";
-    
+
     private final MeterRegistry meterRegistry;
     private final ApplicationProperties config;
     private final ScheduledExecutorService scheduler;
     private final Map<String, Admin> adminClients = new HashMap<>();
 
     @PostConstruct
-    public void init() throws InterruptedException, ExecutionException, TimeoutException {
+    public void init() {
         config.getClusters().forEach(clusterConfig -> {
             Properties props = new Properties();
             props.put("bootstrap.servers", clusterConfig.getServers());
@@ -59,27 +61,60 @@ public class KafkaLagCollector {
             adminClients.put(clusterConfig.getName(), adminClient);
         });
 
+        scheduleJobs();
+    }
+
+    /**
+     * Schedules one or more jobs to collect lag
+     */
+    private void scheduleJobs() {
         for (LagMonitorConfig lagMonitorConfig : config.getLagMonitor().getClusters()) {
-            Admin adminClient = adminClients.get(lagMonitorConfig.getName());
-            List<String> groupIds = null;
-
             if (!lagMonitorConfig.getConsumerGroups().isEmpty()) {
-                groupIds = lagMonitorConfig.getConsumerGroups();
+                for (String groupId : lagMonitorConfig.getConsumerGroups()) {
+                    scheduler.scheduleAtFixedRate(() -> collectConsumerGroupLag(lagMonitorConfig.getName(), groupId), 1,
+                            1, TimeUnit.MINUTES);
+                }
             } else {
-                groupIds = adminClient.listConsumerGroups().valid()
-                        .get(config.getLagMonitor().getTimeoutSeconds(), TimeUnit.SECONDS).stream()
-                        .map(consumerGroup -> consumerGroup.groupId())
-                        .collect(Collectors.toList());
-            }
-
-            for (String groupId : groupIds) {
                 scheduler.scheduleAtFixedRate(() -> {
-                    collectConsumerGroupLag(lagMonitorConfig.getName(), groupId);
+                    collectConsumerGroupLag(lagMonitorConfig);
                 }, 1, 1, TimeUnit.MINUTES);
             }
         }
     }
 
+    /**
+     * Collects the lag for all allowed consumer groups in a cluster
+     * 
+     * @param lagMonitorConfig
+     */
+    private void collectConsumerGroupLag(LagMonitorConfig lagMonitorConfig) {
+        List<Future<?>> futures = new ArrayList<>();
+        Admin adminClient = adminClients.get(lagMonitorConfig.getName());
+
+        try {
+            adminClient.listConsumerGroups().valid()
+                    .get(config.getLagMonitor().getTimeoutSeconds(), TimeUnit.SECONDS).stream()
+                    .filter(group -> isAllowedConsumerGroup(lagMonitorConfig, group.groupId()))
+                    .forEach(group -> futures.add(scheduler
+                            .submit(() -> collectConsumerGroupLag(lagMonitorConfig.getName(), group.groupId()))));
+
+            for (Future<?> future : futures) {
+                future.get();
+            }
+        } catch (ExecutionException | TimeoutException exception) {
+            LOGGER.error("Exception: {}, Message: {}", exception.getClass().getName(),
+                    exception.getMessage());
+        } catch (InterruptedException exception) {
+            Thread.currentThread().interrupt();
+        }
+    }
+
+    /**
+     * Collects the lag for a specific consumer group
+     * 
+     * @param clusterName
+     * @param groupId
+     */
     private void collectConsumerGroupLag(String clusterName, String groupId) {
         LOGGER.debug("Collecting metrics for consumer group '{}'", groupId);
         Admin adminClient = adminClients.get(clusterName);
@@ -133,9 +168,38 @@ public class KafkaLagCollector {
             } else {
                 LOGGER.info("Offsets not found for consumer group '{}'", groupId);
             }
-        } catch (InterruptedException | ExecutionException | TimeoutException e) {
-            LOGGER.error("Group Id: {}, Exception: {}, Message: {}", groupId, e.getClass().getName(),
-                    e.getMessage());
+        } catch (ExecutionException | TimeoutException exception) {
+            LOGGER.error("Group Id: {}, Exception: {}, Message: {}", groupId, exception.getClass().getName(),
+                    exception.getMessage());
+        } catch (InterruptedException exception) {
+            Thread.currentThread().interrupt();
         }
+    }
+
+    /**
+     * Checks if the supplied consumer group name is allowed for monitoring
+     * 
+     * @param config
+     * @param consumerGroup
+     * @return a flag
+     */
+    private boolean isAllowedConsumerGroup(LagMonitorConfig config, String consumerGroup) {
+        boolean isAllowed = false;
+
+        if (config.getGroupAllowlistCompiled().isEmpty()) {
+            if (!config.getGroupDenylistCompiled().isEmpty()) {
+                if (config.getGroupDenylistCompiled().stream()
+                        .noneMatch(pattern -> pattern.matcher(consumerGroup).matches())) {
+                    isAllowed = true;
+                }
+            } else {
+                isAllowed = true;
+            }
+        } else if (config.getGroupAllowlistCompiled().stream()
+                .anyMatch(pattern -> pattern.matcher(consumerGroup).matches())) {
+            isAllowed = true;
+        }
+
+        return isAllowed;
     }
 }