+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## [Unreleased]
 ### Added
 - Initial version. Records kafka lag for specified consumer groups
+- [#4](https://github.com/devatherock/kafka-lag-monitor/issues/4): Accepted regular expressions in the list of consumer groups to monitor
+- [#5](https://github.com/devatherock/kafka-lag-monitor/issues/5): Accept a list of regular expressions for consumer groups to exclude